Rename monotone module. Add two QC tests.
