+++ hgeometry/hgeometry.cabal
@@ -81,6 +81,7 @@ library
                     Data.Geometry.Polygon
                     Data.Geometry.Polygon.Inflate
                     Data.Geometry.Polygon.Convex
+                    Data.Geometry.Polygon.Monotone
 
                     Data.Geometry.BezierSpline
 
@@ -166,8 +167,6 @@ library
 
                     Algorithms.Geometry.SSSP
                     Algorithms.Geometry.SSSP.Naive
-
-                    Algorithms.Geometry.MonotonePolygon.MonotonePolygon
 
                     -- * Embedded Planar Graphs
                     Data.PlaneGraph
